requirements.txt
streamlit
ipython
python-dotenv
#openai==0.28.1
#langchain==0.0.330
matplotlib
scikit-learn
xgboost
shap
pandas
lime
arxiv
chromadb
#pysqlite3-binary 
tiktoken
pypdf
